requirements.txt
flask
ultralytics
torch
Pillow
huggingface_hub